fixed a format bug
git-svn-id: https://jtidy.svn.sourceforge.net/svnroot/jtidy/branches/CodeUpdateAndJava5@962 58265ab0-d70d-0410-8993-b38ad8f1ef24

+++ src/main/java/org/w3c/tidy/AttrCheckImpl.java
@@ -290,7 +290,7 @@ public final class AttrCheckImpl
                 for (int i = 0; i < p.length(); ++i) {
                     char c = p.charAt(i);
                     if ((c > 0x7e) || (c <= 0x20) || (c == '<') || (c == '>')) {
-                        dest.append(String.format("%%%02X", c));
+                        dest.append(String.format("%%%02X", (int) c));
                     } else {
                         dest.append(c);
                     }
